Add full contents of Lab1
The code is capable of loading a dataset of images, finding the
calibration pattern on these images, calibrating the camera
parameters and saving/loading them from a .json file, calculating
a reprojection error and finally saving images with removed
distortion.

The two .json files contain results of camera calibration on data
with and without the use of subpixel pattern coordinates.